docs(memory_match): split modules, apply EXAMPLE_STANDARD, pin cougr-core 1.1.0
Splits the monolithic lib.rs into components.rs (card/board/
game-state components) and systems.rs (reveal/pair-resolution
validation, reset logic) per EXAMPLE_STANDARD.md §3. Rewrites
README.md to cover all eight §4 sections, marks the example
transitional, bumps cougr-core to 1.1.0, and commits Cargo.lock. No
game rule logic was changed.

Refs #223
